Fleshing out data integrity works

+++ CourseTracker.Maui/Data/TermsDB.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.ObjectModel;
-using CourseTracker.Maui.Models;
+﻿using CourseTracker.Maui.Models;
 using CourseTracker.Maui.Services;
 
 namespace CourseTracker.Maui.Data
@@ -7,6 +6,7 @@ namespace CourseTracker.Maui.Data
     public class TermsDB
     {
         SQLiteAsyncConnection _database;
+        CourseDB courseDB;
         public TermsDB()
         {
         }
@@ -49,18 +49,39 @@ namespace CourseTracker.Maui.Data
             await Init();
             if (term.TermId != 0)
             {
-                await _database.UpdateAsync(term);
+                await _database.InsertOrReplaceAsync(term);
             }
             else
             {
-                await _database.InsertAsync(term);
+                await _database.UpdateAsync(term);
             }
         }
         public async Task<int> DeleteTermAsync(Term term)
         {
             await Init();
-            return await _database.DeleteAsync(term);
+            
+            int count = term.CourseCount;
+            var result = await Application.Current.MainPage.DisplayAlert(
+                "WARNING: DELETING TERM", "Are you SURE you want to delete this term?" +
+                " It will also remove " + count + " courses." +
+                "If you wish to keep this data, please move the existing courses to another term first.", "Yes", "No");
+            if (result)
+            {
+                List<Course> allCourses = await courseDB.GetCoursesAsync();
+                foreach (Course course in allCourses)
+                {
+                    if (course.TermId == term.TermId)
+                    {
+                        await courseDB.DeleteCourseAsync(course);                 
+                    }
+                }
+
+              return await _database.DeleteAsync(term);
+            }
+            return 0;
         }
+
+
 
     }
 }
